separate out functionality
