+++ app.js
@@ -11,6 +11,15 @@ app.use(express.json());
 
 const isLoggedIn = (req, res, next)=> {
   if(!req.user){
+    const error = Error('not authorized');
+    error.status = 401;
+    return next(error);
+  }
+  next();
+};
+
+const isAdmin = (req, res, next)=> {
+  if(req.user.role !== 'ADMIN'){
     return next(Error('not authorized'));
   }
   next();
@@ -26,7 +35,11 @@ app.use((req, res, next)=> {
       req.user = auth;
       next();
     })
-    .catch(next);
+    .catch(ex => {
+      const error = Error('not authorized');
+      error.status = 401;
+      next(error);
+    });
 });
 
 app.get('/', (req, res, next)=> res.sendFile(path.join(__dirname, 'index.html')));
@@ -78,24 +91,33 @@ app.delete('/api/removeFromCart/:id', (req, res, next)=> {
     .catch( next );
 });
 
+app.get('/api/products', (req, res, next)=> {
+  db.models.products.read()
+    .then( products => res.send(products))
+    .catch( next );
+});
 
 Object.keys(models).forEach( key => {
-  app.get(`/api/${key}`, (req, res, next)=> {
+  app.get(`/api/${key}`, isLoggedIn, isAdmin, (req, res, next)=> {
     models[key].read({ user: req.user })
       .then( items => res.send(items))
       .catch(next);
   });
-  app.post(`/api/${key}`, (req, res, next)=> {
+  app.post(`/api/${key}`, isLoggedIn, isAdmin, (req, res, next)=> {
     models[key].create({ user: req.user })
       .then( items => res.send(items))
       .catch(next);
   });
 });
 
+app.use((req, res, next)=> {
+  const error = { message: `page not found ${ req.url } for ${ req.method }`, status: 404 };
+  next(error);
+});
+
 app.use((err, req, res, next)=> {
-  console.log(err);
-  res.status(500).send({ message: err.message });
+  console.log(err.status);
+  res.status(err.status || 500).send({ message: err.message });
 });
 
 module.exports = app;
-
