+++ db/models/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  products : require('./products'),
+  users : require('./users'),
+  orders : require('./orders'),
+  lineItems : require('./lineItems')
+};
